// src/api/model/exercise.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */
import type { ExerciseSubTopic } from './exerciseSubTopic';

export interface Exercise {
  answer: string;
  author: string;
  difficulty: number;
  exerciseId: number;
  externalId: string;
  image: string;
  isDone?: boolean;
  isMultipleChoice: boolean;
  subTopic: ExerciseSubTopic;
  tips: string[];
}